+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # RTXNTC Change Log
 
+## 0.6.1 BETA
+
+- Improved the Inference on Feedback mode to add support for standby tiles.
+- Fixed the rendering mode display in the Renderer sample in the reference mode.
+- Implemented handling of `VK_SUBOPTIMAL_KHR` in the SDK apps (https://github.com/NVIDIA-RTX/RTXNTC/issues/3)
+
 ## 0.6.0 BETA
 
 Added the Inference on Feedback mode in the Renderer sample.